+++ packages/xchain-litecoin/src/client.ts
@@ -164,8 +164,7 @@ class Client implements LitecoinClient, XChainClient {
    * @returns {string} The explorer url based on the network.
    */
   getExplorerUrl = (): string => {
-    const networkPath = Utils.isTestnet(this.net) ? '/testnet' : ''
-    return `https://blockstream.info${networkPath}`
+    return Utils.isTestnet(this.net) ? 'https://tltc.bitaps.com' : 'https://ltc.bitaps.com'
   }
 
   /**
@@ -175,7 +174,7 @@ class Client implements LitecoinClient, XChainClient {
    * @returns {string} The explorer url for the given address based on the network.
    */
   getExplorerAddressUrl = (address: Address): string => {
-    return `${this.getExplorerUrl()}/address/${address}`
+    return `${this.getExplorerUrl()}/${address}`
   }
 
   /**
@@ -185,7 +184,7 @@ class Client implements LitecoinClient, XChainClient {
    * @returns {string} The explorer url for the given transaction id based on the network.
    */
   getExplorerTxUrl = (txID: string): string => {
-    return `${this.getExplorerUrl()}/tx/${txID}`
+    return `${this.getExplorerUrl()}/${txID}`
   }
 
   /**