+++ packages/xchain-litecoin/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Uses Bitaps to submit transactions instead of Sochain
 
+### Fix
+
+- Fix `getExplorerUrl` to bitaps.
+
 # v.0.1.0 (2021-02-19)
 
 ### Breaking change